pyproject.toml
[tool.poetry]
name = "fastapi-rest"
version = "0.1.0"
description = ""
authors = ["DarkLorianPrime <lolpokens@mail.ru>"]
readme = "README.md"
packages = [{include = "fastapi_rest"}]

[tool.poetry.dependencies]
python = "^3.11"
databases = "^0.7.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
python-multipart = "^0.0.6"
uvloop = "^0.17.0"
nest-asyncio = "^1.5.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
